chore: moved controller to a separate module
